requirements.txt
discord.py==2.3.2
google-generativeai==0.1.0
python-dotenv==1.0.0
zendesk==2.0.3
